Move data definition headers to src/include/def.
