+++ src/Move.cc
@@ -12,7 +12,7 @@ const array<MovePosition, kMoveNum>& GetAllMove()
 {
   static const array<MovePosition, kMoveNum> all_move_list{
       {
-        #include "MoveDefinition.h"
+        #include "def/MoveDefinition.h"
       }};
 
   return all_move_list;
@@ -22,7 +22,7 @@ const array<MovePosition, kValidMoveNum>& GetAllValidMove()
 {
   static const array<MovePosition, kValidMoveNum> all_valid_move_list{
       {
-        #include "InBoardMoveList.h"
+        #include "def/InBoardMoveList.h"
         kNullMove
       }};
 
@@ -33,7 +33,7 @@ const array<MovePosition, kInBoardMoveNum>& GetAllInBoardMove()
 {
   static const array<MovePosition, kInBoardMoveNum> all_in_board_move_list{
       {
-        #include "InBoardMoveList.h"
+        #include "def/InBoardMoveList.h"
       }};
 
   return all_in_board_move_list;
