+++ src/include/def/ReadBoardPositionLeftDiagonal.h
