+++ src/include/def/ReadBoardPositionRightDiagonal.h
