+++ src/include/def/WriteBoardPositionLeftDiagonal.h
