+++ src/include/def/WriteBoardPositionRightDiagonal.h
